Introduce the e2e testing
