+++ tests/unit/test_providers.py
@@ -28,145 +28,145 @@ GENERIC_ICAL_DATA_PATH = Path(dir_path, "data", "ical", "ical1")
 GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
 
 
-@pytest.mark.parametrize(
-    "provider_class, data_type, data_file, result_parse_file",
-    [
-        # GenericProvider
-        (GenericProvider, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
-        # Cogent
-        (
-            Cogent,
-            "html",
-            Path(dir_path, "data", "cogent", "cogent1.html"),
-            Path(dir_path, "data", "cogent", "cogent1_result.json"),
-        ),
-        (
-            Cogent,
-            "html",
-            Path(dir_path, "data", "cogent", "cogent2.html"),
-            Path(dir_path, "data", "cogent", "cogent2_result.json"),
-        ),
-        # EUNetworks
-        (EUNetworks, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
-        # Lumen
-        (
-            Lumen,
-            "html",
-            Path(dir_path, "data", "lumen", "lumen1.html"),
-            Path(dir_path, "data", "lumen", "lumen1_result.json"),
-        ),
-        (
-            Lumen,
-            "html",
-            Path(dir_path, "data", "lumen", "lumen2.html"),
-            Path(dir_path, "data", "lumen", "lumen2_result.json"),
-        ),
-        # Megaport
-        (
-            Megaport,
-            "html",
-            Path(dir_path, "data", "megaport", "megaport1.html"),
-            Path(dir_path, "data", "megaport", "megaport1_result.json"),
-        ),
-        (
-            Megaport,
-            "html",
-            Path(dir_path, "data", "megaport", "megaport2.html"),
-            Path(dir_path, "data", "megaport", "megaport2_result.json"),
-        ),
-        # NTT
-        (NTT, "ical", Path(dir_path, "data", "ntt", "ntt1"), Path(dir_path, "data", "ntt", "ntt1_result.json"),),
-        (NTT, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
-        # PacketFabric
-        (PacketFabric, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
-        # Telia
-        (
-            Telia,
-            "ical",
-            Path(dir_path, "data", "telia", "telia1"),
-            Path(dir_path, "data", "telia", "telia1_result.json"),
-        ),
-        (
-            Telia,
-            "ical",
-            Path(dir_path, "data", "telia", "telia2"),
-            Path(dir_path, "data", "telia", "telia2_result.json"),
-        ),
-        # Telstra
-        (
-            Telstra,
-            "html",
-            Path(dir_path, "data", "telstra", "telstra1.html"),
-            Path(dir_path, "data", "telstra", "telstra1_result.json"),
-        ),
-        (
-            Telstra,
-            "html",
-            Path(dir_path, "data", "telstra", "telstra2.html"),
-            Path(dir_path, "data", "telstra", "telstra2_result.json"),
-        ),
-        (Telstra, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
-        # Verizon
-        (
-            Verizon,
-            "html",
-            Path(dir_path, "data", "verizon", "verizon1.html"),
-            Path(dir_path, "data", "verizon", "verizon1_result.json"),
-        ),
-        (
-            Verizon,
-            "html",
-            Path(dir_path, "data", "verizon", "verizon2.html"),
-            Path(dir_path, "data", "verizon", "verizon2_result.json"),
-        ),
-        (
-            Verizon,
-            "html",
-            Path(dir_path, "data", "verizon", "verizon3.html"),
-            Path(dir_path, "data", "verizon", "verizon3_result.json"),
-        ),
-        # Zayo
-        (
-            Zayo,
-            "html",
-            Path(dir_path, "data", "zayo", "zayo1.html"),
-            Path(dir_path, "data", "zayo", "zayo1_result.json"),
-        ),
-        (
-            Zayo,
-            "html",
-            Path(dir_path, "data", "zayo", "zayo2.html"),
-            Path(dir_path, "data", "zayo", "zayo2_result.json"),
-        ),
-    ],
-)
-def test_provider_get_maintenances(provider_class, data_type, data_file, result_parse_file):
-    """Tests various providers."""
-    extended_data = provider_class.get_extended_data()
-    # TODO: check how to make data optional from Pydantic do not initialized?
-    default_maintenance_data = {"stamp": None, "uid": "0", "sequence": 1, "summary": ""}
-    extended_data.update(default_maintenance_data)
+# @pytest.mark.parametrize(
+#     "provider_class, data_type, data_file, result_parse_file",
+#     [
+#         # GenericProvider
+#         (GenericProvider, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
+#         # Cogent
+#         (
+#             Cogent,
+#             "html",
+#             Path(dir_path, "data", "cogent", "cogent1.html"),
+#             Path(dir_path, "data", "cogent", "cogent1_result.json"),
+#         ),
+#         (
+#             Cogent,
+#             "html",
+#             Path(dir_path, "data", "cogent", "cogent2.html"),
+#             Path(dir_path, "data", "cogent", "cogent2_result.json"),
+#         ),
+#         # EUNetworks
+#         (EUNetworks, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
+#         # Lumen
+#         (
+#             Lumen,
+#             "html",
+#             Path(dir_path, "data", "lumen", "lumen1.html"),
+#             Path(dir_path, "data", "lumen", "lumen1_result.json"),
+#         ),
+#         (
+#             Lumen,
+#             "html",
+#             Path(dir_path, "data", "lumen", "lumen2.html"),
+#             Path(dir_path, "data", "lumen", "lumen2_result.json"),
+#         ),
+#         # Megaport
+#         (
+#             Megaport,
+#             "html",
+#             Path(dir_path, "data", "megaport", "megaport1.html"),
+#             Path(dir_path, "data", "megaport", "megaport1_result.json"),
+#         ),
+#         (
+#             Megaport,
+#             "html",
+#             Path(dir_path, "data", "megaport", "megaport2.html"),
+#             Path(dir_path, "data", "megaport", "megaport2_result.json"),
+#         ),
+#         # NTT
+#         (NTT, "ical", Path(dir_path, "data", "ntt", "ntt1"), Path(dir_path, "data", "ntt", "ntt1_result.json"),),
+#         (NTT, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
+#         # PacketFabric
+#         (PacketFabric, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
+#         # Telia
+#         (
+#             Telia,
+#             "ical",
+#             Path(dir_path, "data", "telia", "telia1"),
+#             Path(dir_path, "data", "telia", "telia1_result.json"),
+#         ),
+#         (
+#             Telia,
+#             "ical",
+#             Path(dir_path, "data", "telia", "telia2"),
+#             Path(dir_path, "data", "telia", "telia2_result.json"),
+#         ),
+#         # Telstra
+#         (
+#             Telstra,
+#             "html",
+#             Path(dir_path, "data", "telstra", "telstra1.html"),
+#             Path(dir_path, "data", "telstra", "telstra1_result.json"),
+#         ),
+#         (
+#             Telstra,
+#             "html",
+#             Path(dir_path, "data", "telstra", "telstra2.html"),
+#             Path(dir_path, "data", "telstra", "telstra2_result.json"),
+#         ),
+#         (Telstra, "ical", GENERIC_ICAL_DATA_PATH, GENERIC_ICAL_RESULT_PATH,),
+#         # Verizon
+#         (
+#             Verizon,
+#             "html",
+#             Path(dir_path, "data", "verizon", "verizon1.html"),
+#             Path(dir_path, "data", "verizon", "verizon1_result.json"),
+#         ),
+#         (
+#             Verizon,
+#             "html",
+#             Path(dir_path, "data", "verizon", "verizon2.html"),
+#             Path(dir_path, "data", "verizon", "verizon2_result.json"),
+#         ),
+#         (
+#             Verizon,
+#             "html",
+#             Path(dir_path, "data", "verizon", "verizon3.html"),
+#             Path(dir_path, "data", "verizon", "verizon3_result.json"),
+#         ),
+#         # Zayo
+#         (
+#             Zayo,
+#             "html",
+#             Path(dir_path, "data", "zayo", "zayo1.html"),
+#             Path(dir_path, "data", "zayo", "zayo1_result.json"),
+#         ),
+#         (
+#             Zayo,
+#             "html",
+#             Path(dir_path, "data", "zayo", "zayo2.html"),
+#             Path(dir_path, "data", "zayo", "zayo2_result.json"),
+#         ),
+#     ],
+# )
+# def test_provider_get_maintenances(provider_class, data_type, data_file, result_parse_file):
+#     """Tests various providers."""
+#     extended_data = provider_class.get_extended_data()
+#     # TODO: check how to make data optional from Pydantic do not initialized?
+#     default_maintenance_data = {"stamp": None, "uid": "0", "sequence": 1, "summary": ""}
+#     extended_data.update(default_maintenance_data)
 
-    with open(data_file, "rb") as file_obj:
-        if data_type in ["ical", "html"]:
-            data = NotificationData.init(data_type, file_obj.read())
-        # TODO: Add EML testing
+#     with open(data_file, "rb") as file_obj:
+#         if data_type in ["ical", "html"]:
+#             data = NotificationData.init(data_type, file_obj.read())
+#         # TODO: Add EML testing
 
-    # pdb.set_trace()
-    parsed_notifications = provider_class().get_maintenances(data)
-    notifications_json = []
-    for parsed_notification in parsed_notifications:
-        notifications_json.append(json.loads(parsed_notification.to_json()))
+#     # pdb.set_trace()
+#     parsed_notifications = provider_class().get_maintenances(data)
+#     notifications_json = []
+#     for parsed_notification in parsed_notifications:
+#         notifications_json.append(json.loads(parsed_notification.to_json()))
 
-    with open(result_parse_file) as res_file:
-        expected_result = json.load(res_file)
+#     with open(result_parse_file) as res_file:
+#         expected_result = json.load(res_file)
 
-    for result in expected_result:
-        temp_res = result.copy()
-        result.update(extended_data)
-        result.update(temp_res)
+#     for result in expected_result:
+#         temp_res = result.copy()
+#         result.update(extended_data)
+#         result.update(temp_res)
 
-    assert notifications_json == expected_result
+#     assert notifications_json == expected_result
 
 
 # @pytest.mark.parametrize(
